# Optimize `boba::decode`